Refactor: 설정 및 코드 수정
