+++ packages/common-ui/components/atoms/Dropdown.module.scss
@@ -1,2 +1,14 @@
 .dropdownContainer {
+    display: flex;
+    align-items: center;
+}
+
+.spacer {
+    display: inline-block;
+    width: 10px;
+}
+
+.tab {
+    display: inline-block;
+    width: 40px;
 }
